requirements.txt
streamlit
requests
PyPDF2
langchain
langchain-openai
faiss-cpu
crewai
pydantic
openai
fastapi
uvicorn
python-multipart
python-dotenv
pysqlite3-binary>=0.5.2 
langchain-community
fitz
frontend
tools
llama-parse
pymupdf
